+++ pages/detail/index.tsx
@@ -11,8 +11,8 @@ import FeedPanel from '@components/page/meetingDetail/Feed/FeedPanel';
 import { Fragment, useState } from 'react';
 import dayjs from 'dayjs';
 import 'dayjs/locale/ko';
-import FeedCreateModal from '@components/page/meetingDetail/Feed/Modal/FeedModal/FeedCreateModal';
 import { useOverlay } from '@hooks/useOverlay/Index';
+import FeedCreateModal from '@components/page/meetingDetail/Feed/Modal/FeedCreateModal';
 dayjs.locale('ko');
 
 const enum SelectedTab {
